fix: harden remote scrollback preservation
